Fix for missed method get_instance
