--- src/Acr.UserDialogs/Platforms/ios/UserDialogs.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using UIKit;
-
-
-namespace Acr.UserDialogs.Platforms.ios
-{
-    public static partial class UserDialogs
-    {
-        /// <summary>
-        /// OPTIONAL: Initialize iOS user dialogs with your top viewcontroll function
-        /// </summary>
-        public static void Init(Func<UIViewController> viewControllerFunc)
-        {
-            Instance = new UserDialogsImpl(viewControllerFunc);
-        }
-
-
-        static IUserDialogs currentInstance;
-        public static IUserDialogs Instance
-        {
-            get
-            {
-                currentInstance = currentInstance ?? new UserDialogsImpl();
-                return currentInstance;
-            }
-            set => currentInstance = value;
-        }
-    }
-}
